requirements.txt
fastapi
pydantic
boto3
pyjwt
requests
Authlib
python-dotenv
starlette